requirements.txt
# Framework Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-WTF==1.1.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Base de données
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7  # Pour PostgreSQL en production

# Traitement PDF et documents
PyMuPDF==1.23.5
pypdf==3.15.4
python-docx==0.8.11
pytesseract==0.3.10

# IA et ML - Tes dépendances existantes
langchain==0.0.350
langchain-community==0.0.8
langchain-ollama==0.0.1
# langchain-chroma==0.1.2
chromadb==0.4.15
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3

# Traitement d'images et vision
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2

# Audio et synthèse vocale
edge-tts==6.1.8
pydub==0.25.1
speechrecognition==3.10.0

# Génération de documents
reportlab==4.0.4
python-pptx==0.6.22

# Détection de langue et traduction
langdetect==1.0.9

# Utilitaires
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
tqdm==4.66.1
aiofiles==23.2.1
asyncio==3.4.3

# Validation et sécurité
WTForms==3.0.1
email-validator==2.0.0
bcrypt==4.0.1
werkzeug==2.3.7

# Interface web
Jinja2==3.1.2
MarkupSafe==2.1.3

# Développement et tests
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0
redis==5.0.0  # Pour le cache en production

# Utilitaires système
pathlib==1.0.1
uuid==1.30
datetime==5.2
tempfile==1.0
io==1.0
base64==2.0
hashlib==20081119
logging==0.4.9.6
pickle==4.0
shutil==1.0
glob2==0.7
concurrent-futures==3.1.1